Lesson 193 - TD: Using the Form State
submit button's disable attribute is controlled by form's validity.

So if it isn't valid we need to disable submit button.

We get this information by adding a
#f="ngForm"
statement to the <form> to get its infos

next we used this:
 input.ng-invalid.ng-touched
this means
get all the tags that has .ng-invalid and ng-touched classes at the same time.
Then sort them where the tag is input.

In the end it means tag=input classes: .ng-invalid .ng-touched